fix: Endpoint fix

+++ script/trainModel.js
@@ -1,4 +1,4 @@
-import * as tf from '@tensorflow/tfjs';
+import * as tf from '@tensorflow/tfjs-node';
 import fs from 'fs';
 
 const moodMap = { 'Frustrated': 0, 'Sad': 1, 'Anxious': 2, 'Neutral': 3, 'Grateful': 4, 'Happy': 5 };
@@ -17,7 +17,8 @@ const inputs = rawData.map(d => [
 const labels = rawData.map(d => moodMap[d.moodScore]);
 
 const xs = tf.tensor2d(inputs);
-const ys = tf.tensor1d(labels, 'int32');
+// const ys = tf.tensor1d(labels, 'int32');
+const ys = tf.tensor1d(labels, 'float32');
 
 const model = tf.sequential();
 model.add(tf.layers.dense({ inputShape: [5], units: 32, activation: 'relu' }));
